Tests/LibWeb/Text/expected/abspos-escaping-baseline-shifted-inline-block.txt

short offset: 250,80
abs offset: 20,0
abs position matches short position: false
